Reparacion de la advertencia

+++ src/main.cpp
@@ -85,7 +85,7 @@ int main() {
         // Generar bolas cada cierto tiempo
         if (relojGeneracion.getElapsedTime().asSeconds() >= intervaloGeneracion) {
             // Crear una bola en una posición aleatoria
-            BolaDinamica nuevaBola(mundo, {rand() % 800, 0.0f}, 13.0f, 0.01f, 0.7f, sf::Color::White);
+            BolaDinamica nuevaBola(mundo, {static_cast<float>(rand() % 800), 0.0f}, 13.0f, 0.01f, 0.7f, sf::Color::White);
             nuevaBola.setTexture(bolaTexture); // Aplicar textura a la bola
             bolas.push_back(nuevaBola);
             relojGeneracion.restart();
